+++ test/iam-instance-profile/main.tf
@@ -21,10 +21,13 @@ module "instance-profile-s3-full-access" {
   source = "../.."
 
   # name of the role
-  name = "S3FullAccess"
+  name = "S3 Full Access"
 
   # name of the instance profile
-  instance_profile_name = "S3FullAcess"
+  instance_profile_name = "S3 Full Access"
+
+  # name of the policy
+  policy_name = "S3 Full Access"
 
   # the policy granting access
   policy_statements = [