+++ variables.tf
@@ -44,13 +44,13 @@ variable "assume_role_conditions" {
 
 variable "name" {
   type        = string
-  description = "(Optional, Forces new resource) The name of the role. If omitted, Terraform will assign a random, unique name."
+  description = "(Optional, Forces new resource) The name of the role. Invalid characters will be replaced with dashes. If omitted, Terraform will assign a random, unique name."
   default     = null
 }
 
 variable "name_prefix" {
   type        = string
-  description = "(Optional, Forces new resource) Creates a unique name beginning with the specified prefix. Conflicts with name."
+  description = "(Optional, Forces new resource) Creates a unique name beginning with the specified prefix. Invalid characters will be replaced with dashes. Conflicts with name."
   default     = null
 }
 
@@ -100,13 +100,13 @@ variable "policy_statements" {
 
 variable "policy_name" {
   type        = string
-  description = "(Optional) The name of the role policy. If omitted, Terraform will assign a random, unique name."
+  description = "(Optional) The name of the role policy. Invalid characters will be replaced with dashes. If omitted, Terraform will assign a random, unique name."
   default     = null
 }
 
 variable "policy_name_prefix" {
   type        = string
-  description = "(Optional) Creates a unique name beginning with the specified prefix. Conflicts with name."
+  description = "(Optional) Creates a unique name beginning with the specified prefix. Invalid characters will be replaced with dashes. Conflicts with name."
   default     = null
 }
 
@@ -128,13 +128,13 @@ variable "create_instance_profile" {
 
 variable "instance_profile_name" {
   type        = string
-  description = "(Optional, Forces new resource) The profile's name. If omitted, Terraform will assign a random, unique name."
+  description = "(Optional, Forces new resource) The profile's name. Invalid characters will be replaced with dashes. If omitted, Terraform will assign a random, unique name."
   default     = null
 }
 
 variable "instance_profile_name_prefix" {
   type        = string
-  description = "(Optional, Forces new resource) Creates a unique name beginning with the specified prefix. Conflicts with name."
+  description = "(Optional, Forces new resource) Creates a unique name beginning with the specified prefix. Invalid characters will be replaced with dashes. Conflicts with name."
   default     = null
 }